+++ brainscore_vision/model_helpers/activations/temporal/core/inferencer/__init__.py
@@ -1,0 +1,2 @@
+from .base import Inferencer
+from .video import *
